+++ src/Models/Ban.php
@@ -9,30 +9,30 @@
  * file that was distributed with this source code.
  */
 
-namespace Cog\Ban\Models;
+namespace Cog\Laravel\Ban\Models;
 
 use Carbon\Carbon;
-use Cog\Ban\Contracts\Ban as BanContract;
-use Cog\Ownership\Traits\HasMorphOwner;
+use Cog\Contracts\Ban\Ban as BanContract;
+use Cog\Contracts\Ban\Bannable as BannableContract;
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
 /**
  * Class Ban.
  *
- * @package Cog\Ban\Models
+ * @package Cog\Laravel\Ban\Models
  */
 class Ban extends Model implements BanContract
 {
-    use HasMorphOwner,
-        SoftDeletes;
+    use SoftDeletes;
 
     /**
      * The table associated with the model.
      *
      * @var string
      */
-    protected $table = 'ban';
+    protected $table = 'bans';
 
     /**
      * The attributes that are mass assignable.
@@ -54,16 +54,6 @@ class Ban extends Model implements BanContract
     ];
 
     /**
-     * Entity responsible for ban.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphTo
-     */
-    public function createdBy()
-    {
-        return $this->morphTo('created_by');
-    }
-
-    /**
      * Expired timestamp mutator.
      *
      * @param \Carbon\Carbon|string $value
@@ -75,5 +65,40 @@ class Ban extends Model implements BanContract
         }
 
         $this->attributes['expired_at'] = $value;
+    }
+
+    /**
+     * Entity responsible for ban.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\MorphTo
+     */
+    public function createdBy()
+    {
+        return $this->morphTo('created_by');
+    }
+
+    /**
+     * Bannable model.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\MorphTo
+     */
+    public function bannable()
+    {
+        return $this->morphTo('bannable');
+    }
+
+    /**
+     * Scope a query to only include models by owner.
+     *
+     * @param \Illuminate\Database\Eloquent\Builder $query
+     * @param \Cog\Contracts\Ban\Bannable $bannable
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public function scopeWhereBannable(Builder $query, BannableContract $bannable)
+    {
+        return $query->where([
+            'bannable_id' => $bannable->getKey(),
+            'bannable_type' => $bannable->getMorphClass(),
+        ]);
     }
 }