+++ tests/TestCase.php
@@ -9,32 +9,19 @@
  * file that was distributed with this source code.
  */
 
-namespace Cog\Ban\Tests;
+namespace Cog\Tests\Laravel\Ban;
 
-use Cog\Ban\Tests\Stubs\Models\User;
+use Cog\Tests\Laravel\Ban\Stubs\Models\User;
 use Illuminate\Support\Facades\File;
 use Orchestra\Testbench\TestCase as Orchestra;
 
 /**
  * Class TestCase.
  *
- * @package Cog\Ban\Tests
+ * @package Cog\Tests\Laravel\Ban
  */
 abstract class TestCase extends Orchestra
 {
-    /**
-     * Register a callback to be run before the application is destroyed.
-     * TODO: Remove it when will be pushed to Orchestra Testbench package.
-     *
-     * @param  callable  $callback
-     *
-     * @return void
-     */
-    protected function beforeApplicationDestroyed(callable $callback)
-    {
-        array_unshift($this->beforeApplicationDestroyedCallbacks, $callback);
-    }
-
     /**
      * Actions to be performed on PHPUnit start.
      */
@@ -69,7 +56,7 @@ abstract class TestCase extends Orchestra
     protected function getPackageProviders($app)
     {
         return [
-            \Cog\Ban\Providers\BanServiceProvider::class,
+            \Cog\Laravel\Ban\Providers\BanServiceProvider::class,
             \Orchestra\Database\ConsoleServiceProvider::class,
         ];
     }
@@ -87,7 +74,7 @@ abstract class TestCase extends Orchestra
      */
     protected function destroyPackageMigrations()
     {
-        File::cleanDirectory('vendor/orchestra/testbench/fixture/database/migrations');
+        File::cleanDirectory('vendor/orchestra/testbench-core/fixture/database/migrations');
     }
 
     /**