+++ tests/Unit/Traits/HasBannedAtHelpersTest.php
@@ -9,19 +9,18 @@
  * file that was distributed with this source code.
  */
 
-namespace Cog\Ban\Tests\Unit\Traits;
+namespace Cog\Tests\Laravel\Ban\Unit\Traits;
 
 use Carbon\Carbon;
-use Cog\Ban\Models\Ban;
-use Cog\Ban\Tests\Stubs\Models\User;
-use Cog\Ban\Tests\TestCase;
+use Cog\Tests\Laravel\Ban\Stubs\Models\User;
+use Cog\Tests\Laravel\Ban\TestCase;
 
 /**
- * Class HasBannedAtTest.
+ * Class HasBannedAtHelpersTest.
  *
- * @package Cog\Ban\Tests\Unit\Traits
+ * @package Cog\Tests\Laravel\Ban\Unit\Traits
  */
-class HasBannedAtTest extends TestCase
+class HasBannedAtHelpersTest extends TestCase
 {
     /** @test */
     public function it_can_set_banned_flag()