+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to `laravel-ban` will be documented in this file.
 
+## [3.0.0] - 2017-08-27
+
+### Changed
+
+- `Cog\Ban\Contracts\Ban` moved to `Cog\Contracts\Ban\Ban`
+- `Cog\Ban\Contracts\HasBans` moved to `Cog\Contracts\Ban\HasBans`
+- `Cog\Ban\Contracts\BanService` moved to `Cog\Contracts\Ban\BanService`
+- All classes namespaces changed from `Cog\Ban\*` to `Cog\Laravel\Ban\*`
+- Renamed database table `ban` to `bans`
+- Renamed database column `owned_by_id` to `bannable_id`
+- Renamed database column `owned_by_type` to `bannable_type`
+- Renamed trait `HasBans` to `Bannable`
+- Renamed contract `HasBans` to `Bannable`
+- Renamed `Ban::whereOwnedBy($bannable)` to `Ban::whereBannable($bannable)`
+- Renamed Ban model relation `ownedBy` to `bannable`
+
+### Removed
+
+- Dependency Laravel Ownership
+- Removed Ban model `owner` method
+- Removed Ban model `getOwner` method
+
 ## [2.1.0] - 2017-03-21
 
 ### Added
@@ -10,7 +32,7 @@ All notable changes to `laravel-ban` will be documented in this file.
 
 ### Changed
 
-- `HasBans` is a collection of traits `HasBannedAtHelpers`, `HasBannedAtScope`, `HasBansRelation` now.
+- `HasBans` is a collection of traits `HasBannedAtHelpers`, `HasBannedAtScope`, `HasBansRelation` now
 
 ## [2.0.1] - 2017-03-19
 
@@ -29,6 +51,7 @@ All notable changes to `laravel-ban` will be documented in this file.
 
 - Initial release
 
+[3.0.0]: https://github.com/cybercog/laravel-ban/compare/2.1.1...3.0.0
 [2.1.0]: https://github.com/cybercog/laravel-ban/compare/2.0.1...2.1.0
 [2.0.1]: https://github.com/cybercog/laravel-ban/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/cybercog/laravel-ban/compare/1.0.0...2.0.0